Setting: editorconfig
EditorConfig helps maintain consistent coding styles for multiple
developers working on the same project across various editors and IDEs.
The EditorConfig project consists of a file format for defining coding
styles and a collection of text editor plugins that enable editors to
read the file format and adhere to defined styles.

https://editorconfig.org/

This change explicitly enables neovim to use .editorconfig files. This
is already enabled by default.