Plugin: local-highlight: Add
The local-highlight plugin highlights other instances of the word under
the cursor. This happens only on lines currently on the screen. The
intention is that highlighting happens very fast even on large files. I
have also made a custom highlighting group that can be controlled within
the plugin config file.