Plugin: Oil: Add
This is a file explorer plugin that is unique in that files and
directories are edited within a vim buffer using traditional vim
commands. For example, a new file is created by hitting 'o' to create a
new line, typing in the file name, and saving the buffer.

The plugin was added and is configure to lazy load on a keymap. This
plugin also co-opts a few methods of triggering netrw. However, these
don't work until after the plugin has been lazy loaded by the keymap -
this likely needs to be fixed.