Plugin: lazygit: Use plenary window
Using the plenary plugin to create the lazygit window makes the border a
little nicer. This change sets the config options to use the plenary
window creation function and sets the background transparency to zero.
The lazygit.nvim plugin didn't respect the transparency config option
when used with the plenary option. I forked the plugin repo and added
the functionality, hence the change in the plugin source.