Plugin: lspconfig: Dynamic LSP install
Make all LSPs install dynamically based on the presence of a file for
the language in the lang/ folder. The goal of this is to reduce the
number of locations needed to change settings for a given language. The
language file will control all language related plugins and configs
including LSP, DAP, and highlighting settings.