Lang: markdown: Add LSP and grip
This adds a language file for markdown. It installs the LSP as well as a
plugin to add grip control functionality from vim. Grip is a python
package for rendering markdown files as html. It locally hosts a
web-server to view the rendered html in a browser and can export the
html as well. Grip is convenient for editing markdown because it
automatically refreshes the webpage after the corresponding markdown
file has been saved.

I tried to make the vim-grip plugin lazy load. This commit only
implements lazy loading on opening a markdown file. I could not get lazy
loading to work on the plugin's commands. I think this is because there
are issues with how lazy.nvim handles pure vimscript plugins.

I have added a keymap to open the current buffer in grip. This will only
work if the current buffer is a markdown file. This should probably be
improved to not run if it isn't a valid situation. Also, I probably
would want toggle functionality instead of on/off functionality.