Plugin: lazygit: Change keymap initialization
I was already effectively lazy-loading on a keymap anyway since the
keymap was calling the lazy-loading command. I want to get rid of the
Plugin.init() function because some of the plugin's global variables
get assigned default values when the plugin loads. The init function
always happens at neovim startup so some variables will be overwritten.
The better place to assign these global variables would be in
Plugin.config() that runs only when the plugin loads.