Plugin: lualine: Add macro recording
Indicate which macro (letter) is being recorded on the status line. This
makes the 'recording' message show up on the lualine and in the command
line. Because of this, I removed the command line (cmdheight=0). It
still shows up when needed.

Also added some descriptive comments. The laststatus option has not been
changed from default but I placed it here due to its relevance.