Plugin: lualine: highlight showcmd and selectioncount
Make the color of the showcmd and selectioncount text match color of the
lualine's vim mode highlight group. I am manually setting the color RGB
values, but this could likely be done dynamically to match whatever the
theme option is. That would be preferred so that changing the
colorscheme also adjusts the highlighting.