Plugin: overseer: add
Add the overseer plugin which allows for common commands to be run from
inside neovim and provides an interface for viewing and interacting with
them. Many build commands like make are automatically generated and
supported. Custom commands can be specified in overseer's config or per
project in `.vscode/tasks.json`.