Plugin: lint: remove dependency
Remove mason-nvim-lint plugin. It causes issues when trying to use
linters supported by nvim-lint that aren't in Mason's registry. For
example, nvim-lint natively support checkpatch as a linter but it isn't
installed by Mason. In this case, mason-nvim-lint throws an error if it
is passed checkpatch. I know how to manually install Mason packages so
it seemed easier to do it myself to exactly control what gets installed.