Plugin: lspconfig: prevent error
This prevents an error that occurs when a Mason package is installed but
specified anywhere in this neovim config. The line in this change gets
called for all Mason packages. If their opts haven't been specified
somewhere in the config, then it passes `nill` to the setup function -
throwing an error. This change stops nill from being passed. This is
relevant if a Mason package is installed directly via the Mason menu or
if a config file installing a Mason package is deleted.