Plugin: oil: Fix vcs icon config
An error was being thrown by the oil-vcs-status dependency due to icons
not being assigned in the format it wanted. The change fixes that.

I also changed the capitalization of the git icon references in the
utils file. This made things slightly easier and fit the format of other
icon references.