Plugin: tint: Disable by default
This disables lazy-loading on the WinEnter event, effectively disabling
the plugin until the tint commands (created in this file) are executed.
There is a problem where executing the TintToggle command will lazy-load
tint, but will appear to not have an effect. Tint will be setup which
enables tinting, then the toggle command (which triggered the setup)
will turn tinting off. The desired behavior would be TintToggle, loading
the plugin then immediately tinting the screen - the opposite of the
state before the plugin was loaded.