init: add nvim bin folder
This adds a bin foler to the ~/.local/share/nvim directory to hold
programs and packages installed manually by the vim config. The Mason
plugin already does something similar. This is how all of the LSPs and
Linters are installed. I specifically want to install this to have a
place to install fzf in order to have nvim instance of the program that
doesn't interfere with the external version.