Plugin: lspconfig: use default tagfunc
This change stops Neovim from using its own LSP version of tagfunc -
"v:lua.vim.lsp.tagfunc". This tries to use the LSP to find tags instead
of the default vim way of using the tags file. This is a problem when
the LSP is active but not generating tags. An example of this is running
clangd without a compile_commands.json file. This situation causes tags
to just not work.

Also, I like the idea of preserving default tag behavior and adding LSP
behavior separately. This gives the user the option to choose which
functionality they want.