Plugin: luasnip: Abstract the LuaSnip declaration
Move the LuaSnip declaration out of cmp.lua so make LuaSnip
completion-engine agnostic. Now LuaSnip is defined so that it can be
utilized by both Blink and Cmp.

LuaSnip is disabled by default, but can be enabled if snippets written
in lua are desired. The completion engines will automatically be
configured to utilize LuaSnip thanks to the Lazy.nvim `specs` variable.

Friendly-snippets is enabled whether or not LuaSnip is enabled.