Plugin: lazy: stop auto reload config
This stops lazy from trying to auto reload the config for existing
neovim processes. This happens if any change is made that affects the
file hash(?) of a file within the nvim config dir. Auto reloading the
config tends to break the existing sessions without actually bringing in
the new config. Also, the processes that tried to reload the config
become hard to kill.