Keymap: lua-ify highlight cword
The <BS> keymap will highlight the cword or visual selection without
jumping to the next instance (the behavrior of asterisks). This change
converts the vimscript to lua as well as improves functionality of the
keybinds.

The normal mode bind now has the `\<` and `\>` special characters to
only highlight instances with the same word boundary (selecting 'test'
will not highlight instances of 'tests'). This matches asterisk more
closely.

The visual mode mapping now correctly handles newlines more multi-line
selections. It takes advantage of the grep_operator util function so
that all visual modes are handles correctly and so that the state or
registers is preserved.