Plugin: oil: Fuzzy find from current dir
Override the fuzzy file-find/grep keybinds from fzf-lua within Oil
buffers. The keymaps will now do their function with the open oil buffer
directory as the top seach directory. This is useful for trying to seach
a project's sub-directory without exiting nvim.

This adds the keybinds if Fzf-Lua is installed and preserves lazy
loading. Adding the oil opts table as a `spec` of an optional spec
allows this but is kinda gross... This syntax is essentially saying:
change the Fzf-lua config if Fzf-lua is already installed. But the
changed Fzf-lua config is just adding Oil as a spec. This essentially
adds Oil as a dependency without loading it automatically - preserving
Oil's lazy loading.